+++ src/templates/bespoke/utils/transition.ts
@@ -37,7 +37,7 @@ export const _testElementAnimation = (
   elm: HTMLElement,
   callback: (ret: MarpTransitionResolvableKeyframeSettings | undefined) => void
 ) => {
-  /* istanbul ignore next */
+  /* c8 ignore next */
   requestAnimationFrame(() => {
     elm.style.animationPlayState = 'running'
     requestAnimationFrame(() => callback(undefined))
@@ -98,8 +98,9 @@ const resolveKeyframeSetting = (keyframe: string) =>
     if (durationVar) {
       setting.defaultDuration = durationVar
     }
-    // TODO: Abuse animatable property with custom-ident if Shared Element
-    // Transition API had supported in cross-browser.
+    // TODO: Consider to abuse animatable property with custom-ident if View
+    // Transition API had supported in cross-browser and `@property` CSS at-rule
+    // was not yet supported in other browsers well
     //
     // else if (
     //   // CSS variable that is setting within the keyframe cannot read in some
@@ -222,6 +223,9 @@ export const resolveAnimationStyles = (
 ): readonly string[] => {
   const rules: string[] = [
     `:root{${publicCSSVar('direction')}:${opts.backward ? -1 : 1};}`,
+
+    // For marp-bespoke-bespoke-inactive plugin
+    ':root:has(.bespoke-marp-inactive){cursor:none;}',
   ]
 
   const resolveDefaultDuration = (